Enable large numbers of dice with many sides
A single roll of n dice roughly follows a gaussian distribution (specifically, it follows a discrete one). Once there are enough dice, we can just use a gaussian distribution to get a reasonable result. For small numbers of dice, this breaks down a bit, and it's probably faster to just repeatedly roll anyway.

For the maths behind `mean` and `standardDeviation`, see [this SO answer](https://math.stackexchange.com/a/406203).

For the maths behind shifting the gaussian distribution to produce the right values, see [this page](https://www.javamex.com/tutorials/random_numbers/gaussian_distribution_2.shtml) under the heading "Using `nextGaussian`()". (side note: why is it so hard to find a good mathematical source for why shifting a gaussian distribution like this works??)